feat/#10 : MainActivity 이동
