+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.YappOfficial"
         tools:targetApi="31">
+
+        <activity
+            android:name="com.yapp.app.official.MainActivity"
+            android:exported="true"
+            android:theme="@style/Theme.YappOfficial">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+
     </application>
 
 </manifest>
