+++ app/src/main/java/com/yapp/app/official/MainActivity.kt
@@ -1,0 +1,21 @@
+package com.yapp.app.official
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.yapp.feature.home.HomeScreen
+import dagger.hilt.android.AndroidEntryPoint
+
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            HomeScreen()
+        }
+    }
+}
+
