+++ feature/home/src/main/AndroidManifest.xml
@@ -1,17 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
-    <application>
-        <activity
-            android:name="com.yapp.feature.home.MainActivity"
-            android:exported="true"
-            android:theme="@style/Theme.YappOfficial">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-
-    </application>
 
 </manifest>
