+++ feature/home/src/main/java/com/yapp/feature/home/HomeScreen.kt
@@ -1,0 +1,19 @@
+package com.yapp.feature.home
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.yapp.core.designsystem.theme.YappTheme
+
+
+@Composable
+fun HomeScreen (
+){
+    YappTheme {
+        Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+
+        }
+    }
+}
+
